requirements.txt
# Data & Math (Crucial to keep numpy below 2.0 for Coqui/Tensorflow)
numpy<2.0.0
pandas
scikit-learn

# AI & Generation
google-generativeai
coqui-tts
textstat
json5

# File Handling & PDFs
pymupdf
fpdf2

# Audio Processing
pydub
soundfile

# OCR
easyocr
pillow

# Web Framework & Security
flask
flask-cors
flask-bcrypt

# Database & Auth
firebase-admin